course-detail

+++ src/main/java/com/sila/modules/course/dto/CourseDetailResponse.java
@@ -1,10 +1,10 @@
 package com.sila.modules.course.dto;
 
 import com.sila.modules.course.Enum.CategoryStatus;
-import com.sila.modules.profile.dto.res.UserResponse;
-import com.sila.modules.video.dto.VideoListResponse;
 import com.sila.modules.course.Enum.CourseStatus;
 import com.sila.modules.course.Enum.LevelStatus;
+import com.sila.modules.profile.dto.res.UserResponse;
+import com.sila.modules.video.dto.VideoListResponse;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -27,7 +27,7 @@ public class CourseDetailResponse {
   private LevelStatus level;
   private CourseStatus status;
   private CategoryStatus category;
-  private Integer rating;
+  private Double rating;
   private Number reviewsCount;
   private Number duration;
   private Number studentsCount;
